Split CSV output tests from renderer
